🚚 restructure client-side
* put all visitor-related files into visitor directory, structured
  similarly to the cb-admin directory
* move top-level assets into shared/assets
* move common components into shared/components

Related to #296
