--- client/src/App.css
@@ -1,360 +1,0 @@
-html,
-body,
-#root {
-  height: 100%;
-}
-
-.Container {
-  font-family: 'Quicksand', sans-serif;
-  display: flex;
-  align-items: center;
-  min-height: 100%;
-}
-
-.Foreground {
-  background-color: white;
-  color: #2a1756;
-  width: 80%;
-  margin: auto;
-  border-radius: 8px;
-  min-height: 65vh;
-}
-
-.Main {
-  padding-top: 40px;
-  padding-bottom: 40px;
-}
-
-.Privacy {
-  background-color: black;
-  background-image: url('/images/stonehenge.png');
-  background-size: cover;
-  color: white;
-  max-width: 400px;
-  border-radius: 0 8px 8px 0;
-  min-height: 65vh;
-}
-
-.Privacy__image {
-  width: 40%;
-  margin: auto;
-  display: block;
-  padding-top: 3%;
-}
-
-.Privacy__title {
-  display: inline-block;
-  padding: 4%;
-  text-align: center;
-}
-
-.Privacy__text {
-  display: inline-block;
-  padding: 4%;
-  font-size: 0.9rem;
-}
-
-.Button {
-  background-color: orange;
-  color: white;
-  font-size: 1em;
-  border: none;
-  border-radius: 5px;
-  box-shadow: 0px 1px 1px 0px rgba(51, 51, 51, 0.5);
-  width: 60%;
-  margin-left: 20%;
-  padding: 15px;
-  margin-bottom: 20px;
-  margin-top: 20px;
-  cursor: pointer;
-}
-
-.ButtonBack {
-  background-color: #24a597;
-  width: 40%;
-  margin-left: 30%;
-}
-
-.Logoutbutton {
-  background-color: #24a597;
-  color: white;
-  font-size: 1em;
-  border: none;
-  border-radius: 5px;
-  box-shadow: 0px 1px 1px 0px rgba(51, 51, 51, 0.5);
-  width: 40%;
-  margin-left: 30%;
-  padding: 1%;
-  margin-bottom: 2%;
-}
-
-.Form__Label {
-  width: 70%;
-  margin-left: 15%;
-  margin-bottom: 20px;
-  display: block;
-}
-
-.Form__Input {
-  background-color: rgba(100, 68, 233, 0.15);
-  border: none;
-  height: 40px;
-  text-indent: 5px;
-  border-radius: 5px;
-  display: block;
-  width: 98%;
-  padding-left: 2%;
-}
-
-.Form__Select {
-  width: 100%;
-  background-color: rgba(100, 68, 233, 0.15);
-  border: none;
-  height: 40px;
-  text-indent: 5px;
-}
-
-.Video {
-  display: block;
-  width: 80%;
-  object-fit: contain;
-  margin: auto;
-}
-
-.ErrorText {
-  color: red;
-  padding: 2%;
-  text-align: center;
-}
-
-.SuccessText {
-  color: green;
-  padding: 2%;
-}
-
-.hidden {
-  display: none;
-}
-
-h1 {
-  text-align: center;
-  padding-bottom: 20px;
-  margin: 0;
-  padding-top: 50px;
-}
-
-h2 {
-  text-align: center;
-}
-
-.Policy {
-  text-align: center;
-  display: inline-block;
-  padding: 4%;
-  font-size: 0.9rem;
-}
-
-table {
-  text-align: center;
-  margin: auto;
-}
-
-th {
-  padding-bottom: 5%;
-}
-
-.QR__image {
-  margin: auto;
-  display: block;
-}
-
-.capitalise {
-  text-transform: capitalize;
-}
-
-.dashed {
-  border: 1px dashed #ccc;
-}
-
-.visible-printer {
-  display: none;
-}
-
-@media print {
-  .hidden-printer {
-    display: none;
-  }
-  .visible-printer {
-    display: block;
-  }
-  body {
-    background: white;
-  }
-  .Foreground {
-    box-shadow: none;
-    height: auto;
-    width: auto;
-    overflow: visible;
-  }
-  .qr-code-to-print {
-    text-align: center;
-    color: #333;
-  }
-  .qr-code-to-print img {
-    display: inline-block;
-    margin: 0 20px;
-    vertical-align: middle;
-  }
-  .qr-code-to-print .dashed {
-    display: inline-block;
-    padding: 10px 20px;
-  }
-  .qr-code-to-print h5 {
-    margin: 8px 0 12px 0;
-  }
-}
-
-/*Col system from bootstrap 4 - https://github.com/m-spyratos/bootstrap-4-grid/blob/master/css/grid.css*/
-.row {
-  display: flex;
-  flex-direction: row;
-  /*align-items: center;*/
-  align-items: stretch; /* align items in Cross Axis
-  align-content: stretch; /* Extra space in Cross Axis */
-}
-.col-auto {
-  flex: 0 0 auto;
-  width: auto;
-  max-width: none;
-}
-
-.col-1 {
-  flex: 0 0 8.3333333333%;
-  max-width: 8.3333333333%;
-}
-
-.col-2 {
-  flex: 0 0 16.6666666667%;
-  max-width: 16.6666666667%;
-}
-
-.col-3 {
-  flex: 0 0 25%;
-  max-width: 25%;
-}
-
-.col-4 {
-  flex: 0 0 33.3333333333%;
-  max-width: 33.3333333333%;
-}
-
-.col-5 {
-  flex: 0 0 41.6666666667%;
-  max-width: 41.6666666667%;
-}
-
-.col-6 {
-  flex: 0 0 50%;
-  max-width: 50%;
-}
-
-.col-7 {
-  flex: 0 0 58.3333333333%;
-  max-width: 58.3333333333%;
-}
-
-.col-8 {
-  flex: 0 0 66.6666666667%;
-  max-width: 66.6666666667%;
-}
-
-.col-9 {
-  flex: 0 0 75%;
-  max-width: 75%;
-}
-
-.col-10 {
-  flex: 0 0 83.3333333333%;
-  max-width: 83.3333333333%;
-}
-
-.col-11 {
-  flex: 0 0 91.6666666667%;
-  max-width: 91.6666666667%;
-}
-
-.col-12 {
-  flex: 0 0 100%;
-  max-width: 100%;
-}
-
-/*Styling for Activities Admin Page*/
-
-.Activities {
-  padding-top: 20px;
-  display: inline-block;
-  text-align: left;
-  width: 80%;
-  margin: 0 10%;
-}
-
-.Checkbox__Div {
-  display: table;
-  float: left;
-  margin-right: 30px;
-}
-
-#Table__Row {
-  display: table-row;
-}
-
-ul {
-  padding: 0px;
-}
-
-li {
-  list-style-type: none;
-}
-
-li span.delete-item a {
-  text-decoration: none;
-  color: red;
-}
-
-button {
-  border-style: none;
-}
-/*For Pie Chart*/
-canvas {
-  display: block;
-  margin: auto;
-}
-.Dropzone {
-  text-align: center;
-}
-.Dropzone > div {
-  display: inline-block;
-}
-.Dropzone > img {
-  display: block;
-  margin: auto;
-  padding: 20px;
-  width: 200px;
-  height: 200px;
-}
-
-.CB__Logo {
-  width: 150px;
-  height: 150px;
-}
-
-#visitChart {
-  text-align: center;
-}
-
-#barChartDiv {
-  padding-left: 15%;
-  height: 300px;
-  width: 700px;
-}
-
