+++ client/src/shared/assets/images/qrcodelogo.png
