+++ client/src/shared/components/hoc/redirect_after_timeout.js
