+++ client/src/visitors/components/qrprivacy.js
