+++ client/src/visitors/pages/main.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import p2cLogo from '../../qrcodelogo.png';
-import SignupForm from '../visitors/signup_form';
-import NotFound from '../NotFound';
+import p2cLogo from '../../shared/assets/images/qrcodelogo.png';
+import SignupForm from '../components/signup_form';
+import NotFound from '../../shared/components/NotFound';
 import { CbAdmin, Visitors, ErrorUtils } from '../../api';
 import { Heading, Paragraph, Link } from '../../shared/components/text/base';
 import { PrimaryButton } from '../../shared/components/form/base';
