+++ client/src/visitors/pages/qrcode.js
@@ -3,8 +3,8 @@ import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import styled from 'styled-components';
-import PurposeButton from './purposeButton';
-import QRPrivacy from '../visitors/qrprivacy';
+import PurposeButton from '../components/purposeButton';
+import QRPrivacy from '../components/qrprivacy';
 import { Activities, Visitors } from '../../api';
 import { Heading, Paragraph, Link as HyperLink } from '../../shared/components/text/base';
 import { FlexContainerRow, FlexContainerCol } from '../../shared/components/layout/base';
