+++ client/src/visitors/pages/qrerror.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import QRPrivacy from '../visitors/qrprivacy';
+import QRPrivacy from '../components/qrprivacy';
 
 export default () => (
   <div className="row">
